+++ jspwiki-main/src/main/java/org/apache/wiki/auth/user/UserDatabase.java
@@ -68,6 +68,7 @@ public interface UserDatabase {
      * contain any profiles, this method will return a zero-length array.
      *
      * @return the WikiNames
+     * @throws org.apache.wiki.auth.WikiSecurityException
      */
     Principal[] getWikiNames() throws WikiSecurityException;
 
@@ -92,6 +93,7 @@ public interface UserDatabase {
      *
      * @param index the e-mail address of the desired user profile
      * @return the user profile
+     * @throws org.apache.wiki.auth.NoSuchPrincipalException
      */
     UserProfile findByEmail( String index ) throws NoSuchPrincipalException;
 
@@ -101,6 +103,7 @@ public interface UserDatabase {
      *
      * @param index the login name of the desired user profile
      * @return the user profile
+     * @throws org.apache.wiki.auth.NoSuchPrincipalException
      */
     UserProfile findByLoginName( String index ) throws NoSuchPrincipalException;
 
@@ -110,6 +113,7 @@ public interface UserDatabase {
      *
      * @param uid the unique identifier of the desired user profile
      * @return the user profile
+     * @throws org.apache.wiki.auth.NoSuchPrincipalException
      * @since 2.8
      */
     UserProfile findByUid( String uid ) throws NoSuchPrincipalException;
@@ -120,6 +124,7 @@ public interface UserDatabase {
      *
      * @param index the wiki name of the desired user profile
      * @return the user profile
+     * @throws org.apache.wiki.auth.NoSuchPrincipalException
      */
     UserProfile findByWikiName( String index ) throws NoSuchPrincipalException;
 
@@ -129,15 +134,21 @@ public interface UserDatabase {
      *
      * @param index the fill name of the desired user profile
      * @return the user profile
+     * @throws org.apache.wiki.auth.NoSuchPrincipalException
      */
     UserProfile findByFullName( String index ) throws NoSuchPrincipalException;
 
-    /** Initializes the user database based on values from a Properties object. */
+    /** Initializes the user database based on values from a Properties object.
+     * @param engine
+     * @param props
+     * @throws org.apache.wiki.api.exceptions.NoRequiredPropertyException
+     * @throws org.apache.wiki.auth.WikiSecurityException */
     void initialize( Engine engine, Properties props ) throws NoRequiredPropertyException, WikiSecurityException;
 
     /**
      * Factory method that instantiates a new user profile. The {@link UserProfile#isNew()} method of profiles created using
      * this method should return <code>true</code>.
+     * @return user profile
      */
     UserProfile newProfile();
 
@@ -187,4 +198,12 @@ public interface UserDatabase {
      */
     boolean validatePassword( String loginName, String password );
 
+    /**
+     * validates that the proposed password has not been recently used.
+     * @param loginName
+     * @param password
+     * @return false if the password has been recently used, true otherwise
+     * @since 3.0.0
+     */
+    boolean validatePasswordReuse( final String loginName, final String password );
 }